Todo: extending the list of stop words
We need to start compiling a list of zendesk related words that we
would like to ignore. Examples are words like question or help that
could be found frequently but not add any additional information about
the ticket. This is also a good time to start thinking about words that
do provide information about what is in the ticket.